Email_Code/Appwrite_code/requirements.txt
appwrite
newsapi-python
sendgrid